requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
python-decouple==3.3
requests==2.23.0
robin-stocks==1.0.0.7
urllib3==1.25.9
